+++ src/Stack/Config.hs
@@ -652,7 +652,7 @@ loadBuildConfig mproject maresolver mcompiler = do
         }
 
 -- | Get packages from EnvConfig, downloading and cloning as necessary.
--- If the packages have already been downloaded, this uses a cached value (
+-- If the packages have already been downloaded, this uses a cached value.
 getLocalPackages :: forall env. HasEnvConfig env => RIO env LocalPackages
 getLocalPackages = do
     cacheRef <- view $ envConfigL.to envConfigPackagesRef
@@ -673,7 +673,7 @@ getLocalPackages = do
                               $ C.package
                               $ C.packageDescription gpd
                       in (name, (gpd, loc))
-            deps <- (map wrapGPD . concat)
+            deps <- map wrapGPD . concat
                 <$> mapM (parseMultiCabalFilesIndex root) (bcDependencies bc)
 
             checkDuplicateNames $