+++ src/Stack/Docker.hs
@@ -311,7 +311,7 @@ runContainerAndExit getCmdArgs
              (Files.createSymbolicLink
                  (toFilePathNoTrailingSep sshDir)
                  (toFilePathNoTrailingSep (sandboxHomeDir </> sshRelDir))))
-     containerID <- withWorkingDir (toFilePath projectRoot) $ (trim . decodeUtf8) <$> readDockerProcess
+     containerID <- withWorkingDir (toFilePath projectRoot) $ trim . decodeUtf8 <$> readDockerProcess
        (concat
          [["create"
           ,"--net=host"