+++ test/integration/tests/stackage-3185-ignore-bounds-in-snapshot/Main.hs
@@ -1,0 +1,6 @@
+import StackTest
+
+main :: IO ()
+main = do
+  stackErr ["build", "--stack-yaml", "as-extra-dep.yaml", "--dry-run"]
+  stack ["build", "--stack-yaml", "as-snapshot.yaml", "--dry-run"]